+++ NEWS.md
@@ -1,3 +1,15 @@
+Version 2025.02.05
+------------------
+
+* Added momentary spectrum of synth and fading-average momentary spectrum of echoes to the right of spectrograms.
+
+* Factored out scales used by demo players to `scales.hpp`.
+
+* Added `reset` target to Makefile, it removes `_run_/`.
+
+* Tiny adjustments.
+
+
 Version 2024.12.31_1
 --------------------